Allow the producer to disambiguate between several configurations when matching attributes
This commits adds the ability for the _producer_ to tell what configuration to choose whenever multiple matches
are possible and those configurations contain attributes which are not provided by the consumer. For example,
we can imagine a consumer asking for: `flavor=free`, but a consumer has 2 different variants: one with
`flavor=free, buildType=debug`, and the other with `flavor=free, buildType=release`. In that case, we start
using the _consumer_ attribute schema to get the schema of the extra attribute (here, `buildType`), and
ask it to select a configuration knowing that the consumer didn't express an opinion.

The producer, by default, will return the same set of configurations, which will lead to an error: the consumer
has to choose between them. But by implementing the `selectBestMatch` method appropriately, it is now possible
for the consumer to tell which one to use (in which case it would return a single configuration). The semantics
are up to the strategy: it can select a random one (though for obvious reasons it is preferred to always return
the same configuration), or it can perform a comparison of the configurations to select the best.